Integral control of the hessian scaling factor based on accept rate
A preliminary/test implementation of adaptive control of the scaling
factor for the jump distribution.

Sets a window (e.g., 200 steps); it calculates the acceptance rate
over the last window number of steps. The goal is to maintain the recent
acceptance rate near the target (e.g., 0.3). The integral control
comes in in that it adds up all the cumulative errors over the window,
uses that to get a delta value (i.e., if the error adds up to 0, the
sigma value is unchanged, if it's an overestimate, then the sigma value
is decreased by an amount proportional to the integrated error).

This seems to work except that at least in an initial test it appeared
to cause "ringing", with the acceptance rate and the sigma value
oscillating and criss-crossing each other. I filed it away in this
branch in case it becomes useful later on.