Changed way integral adaptation is calculated
Now the cumulative error is normalized by the window size,
and this is used to set the step size directly, not by
adding it to the previous value of sigma. The old way didn't
make sense because the step size wouldn't be calibrated until
the error went from positive to negative. Of course, as the
magnitude of the integrated error goes down, the step size
should adjust accordingly, even if the sign doesn't change.

The trick will be using the deviation to set the step size directly,
especially since the step size can't be negative. I'm currently
experimenting with normalized to the accept rate target, as this
sets the maximum possible cumulative error.