Working implementation of parallel tempering by MPI
Wrote a sample implementation of fitting 2D gaussians. The
implementation required a number of changes to the core bayessb.MCMC
object as well.

Logging of "interesting variables" into the arrays was moved into the
accept/reject move methods as this made it so that external code driving
the walk wouldn't have to do this explicitly.

Because the MPI-based version instructs each chain to perform N steps at
a time, the estimate method had to be modified (rather, a new method,
estimate_nsteps was created) in order to allow N steps to be run at a
time.

Also, I realized that my implementation of PT had been incorrect in that
I was comparing the posteriors for the different chains independent of
their temperatures. This meant that the posterior probability of some
point x was considered to be the same for a chain at a low temperature
and a chain at a high temperature.