Fixed bug in convergence.py
A previous fix (5c411f083a2660e) had accidentally introduced a new problem in
which the list of chains from the mcmcset (mcmc_set.chains) were passed to the
convergence functions, rather than a list of the position arrays. This
is a patch that reinstantiates the list of position arrays so that the
the algorithm doesn't crash.