Changes to reporters to remove NBD model-specific things
In particular, the reporter on estimation parameters included the list
of nbd_sites and nbd_observables, which aren't applicable for a dye
release model, for example. Similarly, I had to modify the sample_fits
and show_fit_at_position functions to not use mcmc.options.tspan as the
universal timespan, and instead to use the timespan returned by the
updated versions of fit_plotting_function which return both the time
coordinates and the values.

I had to (at least temporarily) remove the part of show_fit_at_position
that plotted all of the observables, because it's not clear how to
approach this when the "fit" being shown is fundamentally a titration,
not a single concentration.