Added abstract chainloader superclass
Idea is that chains will be passed in to Reporters as instances of
ChainLoader, so that they only need to be loaded at report time. This
prevents having to load all the chains into numpy arrays and passing
them in, while allowing for flexibility in the type of chain/sampler
used.